[StaffWikis] Refactor to move to a less awkward URL scheme
