+++ app/javascript/src/styles/views/staff_wikis/_staff_wikis.scss
@@ -1,4 +1,4 @@
-body.c-staff-wikis.a-show #page {
+body.c-staff-staff-wikis.a-show #page {
   padding: 0;
   display: flex;
   flex-flow: column;
