+++ app/views/staff/staff_wiki_versions/diff.html.erb
@@ -10,7 +10,7 @@
   </div>
 </div>
 
-<%= render "staff_wikis/secondary_links" %>
+<%= render "staff/staff_wikis/secondary_links" %>
 
 <% content_for(:page_title) do %>
   Staff Wiki Versions Comparison - <%= @thispage.pretty_title %>
