+++ app/views/staff/staff_wiki_versions/index.html.erb
@@ -12,7 +12,7 @@
   </div>
 </div>
 
-<%= render "staff_wikis/secondary_links" %>
+<%= render "staff/staff_wikis/secondary_links" %>
 
 <% content_for(:page_title) do %>
   Staff Wiki Versions
