+++ app/views/staff/staff_wiki_versions/show.html.erb
@@ -8,7 +8,7 @@
   </div>
 </div>
 
-<%= render "staff_wikis/secondary_links" %>
+<%= render "staff/staff_wikis/secondary_links" %>
 
 <% content_for(:page_title) do %>
   Staff Wiki Version
