+++ app/views/staff/staff_wikis/edit.html.erb
