--- app/views/staff_wikis/search.html.erb
@@ -1,13 +1,0 @@
-<div id="c-staff-wikis">
-  <div id="a-search">
-    <h1>Search Staff Wiki</h1>
-
-    <%= render "search" %>
-  </div>
-</div>
-
-<%= render "secondary_links" %>
-
-<% content_for(:page_title) do %>
-  Search Staff Wiki
-<% end %>
